+++ src/core_write.cpp
@@ -4,7 +4,7 @@
 
 #include "core_io.h"
 #include "univalue/univalue.h"
-#include "script.h"
+#include "scriptutils.h"
 #include "core.h"
 #include "serialize.h"
 #include "util.h"
