+++ src/keystore.cpp
@@ -7,7 +7,7 @@
 
 #include "crypter.h"
 #include "key.h"
-#include "script.h"
+#include "scriptutils.h"
 
 #include <boost/foreach.hpp>
