+++ src/test/DoS_tests.cpp
@@ -12,7 +12,7 @@
 #include "main.h"
 #include "net.h"
 #include "pow.h"
-#include "script.h"
+#include "scriptutils.h"
 #include "serialize.h"
 
 #include <stdint.h>
