+++ src/test/sigopcount_tests.cpp
@@ -3,7 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "key.h"
-#include "script.h"
+#include "scriptutils.h"
 #include "uint256.h"
 
 #include <vector>
