+++ src/test/transaction_tests.cpp
@@ -8,7 +8,7 @@
 #include "key.h"
 #include "keystore.h"
 #include "main.h"
-#include "script.h"
+#include "scriptutils.h"
 #include "core_io.h"
 
 #include <map>
